move util to examples
